Add danger clear button
